play_gameを独立したcrateにしてtigtag_insideへリネームした
