--- tigtag/src/play_game/mod.rs
@@ -1,29 +1,0 @@
-use super::*;
-
-//ゲームロジック
-mod schedule;
-pub use schedule::Schedule;
-
-//play_game内共通
-mod common;
-use common::*;
-
-//UIの処理
-mod header;
-mod ui;
-use ui::*;
-
-//pause処理
-mod pause;
-
-//マップ、自機、追手の処理
-mod map;
-use map::GridToPixelOnMap;
-mod player;
-mod chasers;
-mod detection;
-
-//デモ
-mod demo;
-
-//End of code.
