--- tigtag/src/play_game/ui/mod.rs
@@ -1,12 +1,0 @@
-use super::*;
-
-//UIの表示、振る舞い
-pub mod effect;
-pub use effect::AddOnTraitForText;
-
-pub mod title_demo;
-pub mod stage_start;
-pub mod stage_clear;
-pub mod game_over;
-
-//End of code.
