+++ tigtag_inside/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "tigtag_inside"
+version.workspace = true
+edition.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bevy.workspace = true
+once_cell.workspace = true
+rand.workspace = true
+
+public.workspace = true
