+++ tigtag_inside/src/demo/mod.rs
@@ -1,0 +1,13 @@
+use super::*;
+
+//デモ用スケジュール
+pub mod schedule;
+pub use schedule::Schedule;
+
+//デモ用の自機アルゴリズム
+pub mod auto_drive;
+
+//デモの情報表示
+mod footer;
+
+//End of code.
